Conflicts:
	GestionMagasin-ejb/src/java/facades/FournisseurFacadeLocal.java
	GestionMagasin-war/web/chefDeRayon/ajouterFournisseur.jsp
	GestionMagasin-war/web/chefDeRayon/listeArticle.jsp
	GestionMagasin-war/web/web-inf/web.xml

+++ GestionMagasin-ejb/src/java/facades/RayonFacade.java
@@ -54,7 +54,7 @@ public class RayonFacade extends AbstractFacade<Rayon> implements RayonFacadeLoc
     }
     
     @Override
-    public Rayon rechercherRayonParId(int idRayon) {
+    public Rayon rechercherRayonParId(long idRayon) {
         
     Rayon result;
